+++ lib/nnet-component.h
@@ -9,8 +9,8 @@
 namespace snowboy {
 	struct MatrixBase;
 	class ChunkInfo {
-		int32_t m_feat_dim;
-		int32_t m_num_chunks;
+		size_t m_feat_dim;
+		size_t m_num_chunks;
 		int32_t m_first_offset;
 		int32_t m_last_offset;
 		std::vector<int32_t> m_offsets;
@@ -22,13 +22,13 @@ namespace snowboy {
 			  m_first_offset(0), m_last_offset(0),
 			  m_offsets() {}
 
-		ChunkInfo(int32_t feat_dim, int32_t num_chunks,
+		ChunkInfo(size_t feat_dim, size_t num_chunks,
 				  int32_t first_offset, int32_t last_offset) noexcept
 			: m_feat_dim(feat_dim), m_num_chunks(num_chunks),
 			  m_first_offset(first_offset), m_last_offset(last_offset),
 			  m_offsets() { Check(); }
 
-		ChunkInfo(int32_t feat_dim, int32_t num_chunks,
+		ChunkInfo(size_t feat_dim, size_t num_chunks,
 				  const std::vector<int32_t> offsets)
 			: m_feat_dim(feat_dim), m_num_chunks(num_chunks),
 			  m_first_offset(offsets.front()), m_last_offset(offsets.back()),
@@ -45,9 +45,9 @@ namespace snowboy {
 		void Check() const;
 
 		// Not in snowboy
-		int32_t NumRows() const { return m_num_chunks * ChunkSize(); }
-		int32_t NumCols() const { return m_feat_dim; }
-		int32_t NumChunks() const { return m_num_chunks; }
+		size_t NumRows() const { return m_num_chunks * ChunkSize(); }
+		size_t NumCols() const { return m_feat_dim; }
+		size_t NumChunks() const { return m_num_chunks; }
 		int32_t ChunkSize() const { return !m_offsets.empty() ? m_offsets.size() : (m_last_offset - m_first_offset + 1); }
 		void MakeOffsetsContiguous() {
 			m_offsets.clear();
